pyproject.toml
[project]
name = "faith-dive"
version = "0.1.0"
description = "A Progressive Web App for Bible verse search and journaling"
authors = [
    {name = "mike-maitoza"}
]
readme = "README.md"
requires-python = "^3.12"
dependencies = [
    "fastapi (>=0.118.0,<0.119.0)",
    "uvicorn (>=0.37.0,<0.38.0)",
    "sqlalchemy (>=2.0.43,<3.0.0)",
    "alembic (>=1.16.5,<2.0.0)",
    "python-multipart (>=0.0.20,<0.0.21)",
    "jinja2 (>=3.1.6,<4.0.0)",
    "aiofiles (>=24.1.0,<25.0.0)",
    "requests (>=2.32.5,<3.0.0)",
    "python-dotenv (>=1.1.1,<2.0.0)",
    "pydantic-settings (>=2.11.0,<3.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[dependency-groups]
dev = [
    "pytest (>=8.4.2,<9.0.0)",
    "pytest-asyncio (>=1.2.0,<2.0.0)",
    "httpx (>=0.28.1,<0.29.0)",
    "pytest-cov (>=5.0.0,<6.0.0)",
    "pytest-xdist (>=3.6.0,<4.0.0)",
    "pytest-mock (>=3.14.0,<4.0.0)",
    "pytest-benchmark (>=4.0.0,<5.0.0)",
    "factory-boy (>=3.3.1,<4.0.0)",
    "faker (>=28.4.1,<29.0.0)",
    "responses (>=0.25.3,<1.0.0)",
    "freezegun (>=1.5.1,<2.0.0)",
    "sqlalchemy-utils (>=0.41.2,<1.0.0)",
    "testcontainers (>=4.8.1,<5.0.0)"
]
